Major refactor
Make the FilterModel in ultimate control of everything.  The search
string will be a by-product of what is in filter model, not the other
way around.

move all the logic for grouping and filtering todos to the model level,
rather than component level.

Drastically shrink up the amount of shit that was in the todolist
component, and instead rely on the model to sort things out.

Finally, group together filtering and grouping into the FilterModel, because there is no
reality in which these will be a separate thing.